Move `objc2::ProtocolObject` to `objc2::runtime::ProtocolObject`
