+++ crates/test-assembly/crates/test_extern_protocol/expected/apple-aarch64.s
@@ -14,9 +14,9 @@ Lloh1:
 	.p2align	2
 _dyn_call:
 Lloh2:
-	adrp	x8, L_OBJC_SELECTOR_REFERENCES_36d5bd8a3470dcf9@PAGE
+	adrp	x8, L_OBJC_SELECTOR_REFERENCES_acbdb619e79b01b6@PAGE
 Lloh3:
-	ldr	x1, [x8, L_OBJC_SELECTOR_REFERENCES_36d5bd8a3470dcf9@PAGEOFF]
+	ldr	x1, [x8, L_OBJC_SELECTOR_REFERENCES_acbdb619e79b01b6@PAGEOFF]
 	b	_objc_msgSend
 	.loh AdrpLdr	Lloh2, Lloh3
 
@@ -28,9 +28,9 @@ _dyn_consume:
 	add	x29, sp, #16
 	mov	x19, x0
 Lloh4:
-	adrp	x8, L_OBJC_SELECTOR_REFERENCES_36d5bd8a3470dcf9@PAGE
+	adrp	x8, L_OBJC_SELECTOR_REFERENCES_acbdb619e79b01b6@PAGE
 Lloh5:
-	ldr	x1, [x8, L_OBJC_SELECTOR_REFERENCES_36d5bd8a3470dcf9@PAGEOFF]
+	ldr	x1, [x8, L_OBJC_SELECTOR_REFERENCES_acbdb619e79b01b6@PAGEOFF]
 	bl	_objc_msgSend
 	mov	x0, x19
 	ldp	x29, x30, [sp, #16]
@@ -43,20 +43,20 @@ l_anon.[ID].0:
 	.ascii	"MyProtocol"
 
 	.section	__DATA,__objc_imageinfo,regular,no_dead_strip
-	.globl	L_OBJC_IMAGE_INFO_36d5bd8a3470dcf9
+	.globl	L_OBJC_IMAGE_INFO_acbdb619e79b01b6
 	.p2align	2
-L_OBJC_IMAGE_INFO_36d5bd8a3470dcf9:
+L_OBJC_IMAGE_INFO_acbdb619e79b01b6:
 	.asciz	"\000\000\000\000@\000\000"
 
 	.section	__TEXT,__objc_methname,cstring_literals
-	.globl	L_OBJC_METH_VAR_NAME_36d5bd8a3470dcf9
-L_OBJC_METH_VAR_NAME_36d5bd8a3470dcf9:
+	.globl	L_OBJC_METH_VAR_NAME_acbdb619e79b01b6
+L_OBJC_METH_VAR_NAME_acbdb619e79b01b6:
 	.asciz	"aMethod"
 
 	.section	__DATA,__objc_selrefs,literal_pointers,no_dead_strip
-	.globl	L_OBJC_SELECTOR_REFERENCES_36d5bd8a3470dcf9
+	.globl	L_OBJC_SELECTOR_REFERENCES_acbdb619e79b01b6
 	.p2align	3
-L_OBJC_SELECTOR_REFERENCES_36d5bd8a3470dcf9:
-	.quad	L_OBJC_METH_VAR_NAME_36d5bd8a3470dcf9
+L_OBJC_SELECTOR_REFERENCES_acbdb619e79b01b6:
+	.quad	L_OBJC_METH_VAR_NAME_acbdb619e79b01b6
 
 .subsections_via_symbols
