+++ crates/test-assembly/crates/test_extern_protocol/expected/apple-armv7.s
@@ -15,8 +15,8 @@ LPC0_0:
 	.p2align	2
 	.code	32
 _dyn_call:
-	movw	r1, :lower16:(L_OBJC_SELECTOR_REFERENCES_36d5bd8a3470dcf9-(LPC1_0+8))
-	movt	r1, :upper16:(L_OBJC_SELECTOR_REFERENCES_36d5bd8a3470dcf9-(LPC1_0+8))
+	movw	r1, :lower16:(L_OBJC_SELECTOR_REFERENCES_acbdb619e79b01b6-(LPC1_0+8))
+	movt	r1, :upper16:(L_OBJC_SELECTOR_REFERENCES_acbdb619e79b01b6-(LPC1_0+8))
 LPC1_0:
 	ldr	r1, [pc, r1]
 	b	_objc_msgSend
@@ -27,9 +27,9 @@ LPC1_0:
 _dyn_consume:
 	push	{r4, r7, lr}
 	add	r7, sp, #4
-	movw	r1, :lower16:(L_OBJC_SELECTOR_REFERENCES_36d5bd8a3470dcf9-(LPC2_0+8))
+	movw	r1, :lower16:(L_OBJC_SELECTOR_REFERENCES_acbdb619e79b01b6-(LPC2_0+8))
 	mov	r4, r0
-	movt	r1, :upper16:(L_OBJC_SELECTOR_REFERENCES_36d5bd8a3470dcf9-(LPC2_0+8))
+	movt	r1, :upper16:(L_OBJC_SELECTOR_REFERENCES_acbdb619e79b01b6-(LPC2_0+8))
 LPC2_0:
 	ldr	r1, [pc, r1]
 	bl	_objc_msgSend
@@ -42,20 +42,20 @@ l_anon.[ID].0:
 	.ascii	"MyProtocol"
 
 	.section	__DATA,__objc_imageinfo,regular,no_dead_strip
-	.globl	L_OBJC_IMAGE_INFO_36d5bd8a3470dcf9
+	.globl	L_OBJC_IMAGE_INFO_acbdb619e79b01b6
 	.p2align	2
-L_OBJC_IMAGE_INFO_36d5bd8a3470dcf9:
+L_OBJC_IMAGE_INFO_acbdb619e79b01b6:
 	.asciz	"\000\000\000\000@\000\000"
 
 	.section	__TEXT,__objc_methname,cstring_literals
-	.globl	L_OBJC_METH_VAR_NAME_36d5bd8a3470dcf9
-L_OBJC_METH_VAR_NAME_36d5bd8a3470dcf9:
+	.globl	L_OBJC_METH_VAR_NAME_acbdb619e79b01b6
+L_OBJC_METH_VAR_NAME_acbdb619e79b01b6:
 	.asciz	"aMethod"
 
 	.section	__DATA,__objc_selrefs,literal_pointers,no_dead_strip
-	.globl	L_OBJC_SELECTOR_REFERENCES_36d5bd8a3470dcf9
+	.globl	L_OBJC_SELECTOR_REFERENCES_acbdb619e79b01b6
 	.p2align	2
-L_OBJC_SELECTOR_REFERENCES_36d5bd8a3470dcf9:
-	.long	L_OBJC_METH_VAR_NAME_36d5bd8a3470dcf9
+L_OBJC_SELECTOR_REFERENCES_acbdb619e79b01b6:
+	.long	L_OBJC_METH_VAR_NAME_acbdb619e79b01b6
 
 .subsections_via_symbols
